Rust_coder_llm/Rust_coder_llm/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pymongo
requests